BUG: Fix errors in atom position updates
This mistake in the code (and suggested fix) was kindly pointed out by @gpetretto and is the cause of some of the jittery motions and poor looping that have been observed. It's a classic Python pass-by-reference bug! The result was a sort of poor numerical integration over a sine function - which in turn looks like a sensible oscillation. Users are encouraged to update to this version ASAP. There is a side-effect of changing the scale of oscillations, so the default ranges may need tuning.